+++ frontend/pages/collections/[slug].page.tsx
