+++ frontend/pages/collections/index.page.tsx
@@ -1,0 +1,63 @@
+import axiosWithoutAuth from "@/axiosWithoutAuth";
+import Layout from "@/components/Layout";
+import PageHeader from "@/components/PageHeader";
+import Pagination from "@/components/Pagination";
+import { Collection } from "@/types/modules";
+import { CardContent } from "@mui/material";
+import Grid from "@mui/material/Grid";
+import { GetServerSideProps } from "next";
+import Link from "next/link";
+import { FC } from "react";
+import { Card, Container } from "react-bootstrap";
+
+interface CollectionProps {
+  collectionsData: {
+    results: Collection[];
+    totalPages: number;
+  };
+}
+
+const Collections: FC<CollectionProps> = ({ collectionsData }: CollectionProps) => {
+  //Grid Component for collection card
+  const collections = collectionsData["results"] || [];
+  const collectionCards = collections.map((collection) => (
+    <Grid item key={collection.slug} xs={6} sm={4} md={3}>
+      <Link href={`/collections/${collection.slug}`}>
+        <a>
+          <Card>
+            <CardContent>{collection.title}</CardContent>
+          </Card>
+        </a>
+      </Link>
+    </Grid>
+  ));
+  return (
+    <Layout title={collections}>
+      <PageHeader>Collections</PageHeader>
+      <Pagination count={collectionsData["totalPages"]} />
+      <Container>
+        <Grid container spacing={2}>
+          {collectionCards}
+        </Grid>
+      </Container>
+    </Layout>
+  );
+};
+
+export default Collections;
+
+export const getServerSideProps: GetServerSideProps = async (context) => {
+  let collectionsData = {};
+
+  await axiosWithoutAuth
+    .get("http://django:8000/api/collections/", { params: context.query })
+    .then((response) => {
+      collectionsData = response.data;
+    });
+
+  return {
+    props: {
+      collectionsData,
+    },
+  };
+};
